+++ CHANGELOG.md
@@ -6,8 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- 0395: Support for parameterized Computed Properties.
+- 0345: Asset Rendition Dispatcher framework allowing for named, cacheable renditions.
+
 ### Fixed
-- #0388: Corrected spelling of Boolean in dialog value.
+
+- 0390: Fixed issue where initial values from query parameters were not respected in Search / Sort component's HTL.
+- 0388: Corrected spelling of Boolean in dialog value.
+- 0400: InternalRedirectRenditionDispatcherImpl now supports asset paths with spaces.
 
 ## [v1.7.0]